Release/0.1.14 (#29)
* Pump suspensions on basal chart

* show pump basal profile if no looping

* a manual bolus limited by maxBolus

* Bold glucose delta

* fix bolus calculation

* bump version

+++ FreeAPS/Sources/Models/BloodGlucose.swift
@@ -25,7 +25,7 @@ struct BloodGlucose: JSON, Identifiable, Hashable {
     let direction: Direction?
     let date: Decimal
     let dateString: Date
-    let filtered: Double?
+    let filtered: Decimal?
     let noise: Int?
 
     var glucose: Int?
